categorize extensions as a system table in the sdk
