+++ sdk/src/schema/index.ts
@@ -2,6 +2,7 @@ export * from './activity.js';
 export * from './collection.js';
 export * from './core.js';
 export * from './dashboard.js';
+export * from './extension.js';
 export * from './field.js';
 export * from './file.js';
 export * from './flow.js';
